move new dashboard components to main dashboard
